Update zinit initialization logic and enhance reload function in zshrc.utils; add error handling for authorized keys update in sync.sh

+++ sync.sh
@@ -97,16 +97,21 @@ for source_file in $files; do
     color_echo GREEN "  🔗  Linked: $relative_path"
 done
 
+color_echo BLUE "🔧  Updating authorized keys..."
 # Use github authorized keys to add to ~/.ssh/authorized_keys
-wget https://github.com/agoodkind.keys -O "$HOME"/.ssh/authorized_keys.tmp
-# append missing keys to ~/.ssh/authorized_keys
-touch "$HOME"/.ssh/authorized_keys  # Ensure file exists
-while IFS= read -r key || [ -n "$key" ]; do
-    if ! grep -q "$key" "$HOME"/.ssh/authorized_keys; then
-        echo "$key" >> "$HOME"/.ssh/authorized_keys
-    fi
-done < "$HOME"/.ssh/authorized_keys.tmp
-rm -f "$HOME"/.ssh/authorized_keys.tmp
+if ! wget https://github.com/agoodkind.keys -O "$HOME"/.ssh/authorized_keys.tmp; then
+    color_echo RED "❌  Failed to download authorized keys" && exit 1
+else
+    # append missing keys to ~/.ssh/authorized_keys
+    touch "$HOME"/.ssh/authorized_keys  # Ensure file exists
+    while IFS= read -r key || [ -n "$key" ]; do
+        if ! grep -q "$key" "$HOME"/.ssh/authorized_keys; then
+            echo "$key" >> "$HOME"/.ssh/authorized_keys
+        fi
+    done < "$HOME"/.ssh/authorized_keys.tmp
+    rm -f "$HOME"/.ssh/authorized_keys.tmp
+    color_echo GREEN "  ✅  Authorized keys updated"
+fi
 
 # Symlink all .sh scripts to ~/.local/bin without .sh extension
 color_echo YELLOW "🔗 Linking scripts to ~/.local/bin..."
